Replace lxml with HTMLParser
This change is effectively a complete rewrite of the script, since parsing
the HTML is 95% of what the script does.

The problem with lxml is that it's not part of Python itself, and I've been
unable to compile it on my Mac.  HTMLParser has its problems, but it seems
to work fine with kickstarter.com, and it comes with Python.  Now this
script will run on Macs unmodified.

I also removed optparse, since we don't really support any options anyway.

Signed-off-by: Timur Tabi <timur@tabi.org>