Reset the list of rewards every time we process the web page
Since we're returning a list instead of a dictionary, it's important
that we reset the list every time we process the web page.  This is
because the list is created via the append() method, so we were growing
the list every time we processed the web page.

This bug was introduced in commit "Return a list of tuples instead of an
dictionary of tuples".  Technically, this bug was always there, except
updates to a dictionary replace the existing values.  So if a pledge level
is deleted on the web site, then it won't be deleted from the dictionary.

This bug was discovered while adding support for multiple pledge levels.

Signed-off-by: Timur Tabi <timur@tabi.org>