Update to handle latest web site changes
Kickstarter changed the web site again, and this time significantly.

The web site no longer contains information as to how many pledges are
available (e.g. "1 of 100 remaining!").  So KickstarterHTMLParser now
just returns a list of unavailable pledge levels.  This actually simplifies
things a lot.

The main loop now just checks if the selected pledge(s) exists at all.
If it doesn't, then we know that a previous unavailable pledge level is now
available, and so a web browser window is opened.

Signed-off-by: Timur Tabi <timur@tabi.org>